Splited classes into files again.
