+++ apple_store/OriginalRefurbishedDevice.py
@@ -1,15 +1,9 @@
 #! /usr/bin/python3
 # -*- coding: utf-8 -*-
 
+
 class OriginalRefurbishedDevice:
     ''' Realization of the original refurbished device '''
 
     def __str__(self) -> str:
         return "An original refurbished device"
-
-
-class ReplicaDevice:
-    ''' Realization of the replica from the Odessa basement '''
-
-    def __str__(self) -> str:
-        return "A Replica from the Odessa basement"
