+++ apple_store/ReplicaDevice.py
@@ -1,0 +1,9 @@
+#! /usr/bin/python3
+# -*- coding: utf-8 -*-
+
+
+class ReplicaDevice:
+    ''' Realization of the replica from the Odessa basement '''
+
+    def __str__(self) -> str:
+        return "A Replica from the Odessa basement"
