+++ apple_store/ReplicaDeviceFactory.py
@@ -2,15 +2,7 @@
 # -*- coding: utf-8 -*-
 
 from IDeviceFactory import IDeviceFactory
-from Device import OriginalRefurbishedDevice, ReplicaDevice
-
-
-class OriginalRefurbishedDeviceFactory(IDeviceFactory):
-    ''' A specific factory for original refurbished devices '''
-
-    def create_device(self):
-        # Implementation of creation of original devices
-        return OriginalRefurbishedDevice()
+from ReplicaDevice import ReplicaDevice
 
 
 class ReplicaDeviceFactory(IDeviceFactory):
